Support accessing Connection from multiple event loops
This changes the connection object to stop using the `loop` argument
passed to it at creation in favor of creating futures from whatever
event loop is the current loop at the time of executing a query.
To support Python 3.6, which does not have `Future.get_loop()`, this
adds a simple helper function that will use the `Future._loop` value
when necessary.

This deprecates the `loop` parameter to the Connection constructor
and the `connect()` helper. Deprecation warnings will be generated
when passing a `loop` parameter to either one.